Use define to determine if c++11 or not
So the developer doesn't have to worry about what version of c++ a file
is being compiled as.
